just one request manager
