+++ BeeKit/ServiceLocator.swift
@@ -15,7 +15,7 @@ public class ServiceLocator {
   public static let persistentContainer = BeeminderPersistentContainer.create()
 
   public static let requestManager = RequestManager()
-  public static let signedRequestManager = SignedRequestManager(requestManager: requestManager)
+  
   public static let currentUserManager = CurrentUserManager(
     requestManager: requestManager,
     container: persistentContainer,
